requirements.txt
tensorflow>=2.12
numpy
matplotlib
flask
pillow

